requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.4.0
azure-ai-agents==1.1.0b3
azure-ai-documentintelligence==1.0.2
azure-ai-language-conversations==1.1.0
azure-ai-projects==1.0.0b12
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudevents==1.12.0
colorama==0.4.6
cryptography==45.0.4
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docx2txt==0.9
dotenv==0.9.9
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
lxml==6.0.0
MarkupSafe==3.0.2
more-itertools==10.7.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
narwhals==1.45.0
nest-asyncio==1.6.0
numpy==2.3.1
openai==1.93.0
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
pandas==2.3.0
parse==1.20.2
pathable==0.4.4
pillow==11.3.0
prance==25.4.8.0
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pybars4==0.9.13
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
PyMuPDF==1.26.1
pymupdf4llm==0.0.25
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rpds-py==0.25.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scipy==1.16.0
semantic-kernel==1.34.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.46.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.1
yarl==1.20.1
zipp==3.23.0
